some more refactoring
cleanup of javadoc
